Player view, window design
Adds a PakLevel that manages level loading/unloading.
Adds a PakMan that is going to be our main.
Edited the MouseOrbiter to fix our zoom issue.
Fixed the size/position of the camera in the World panel.
Reviewed the design of the app (WIP!!!)
Removed some useless tasks for our use in simbad internal API.
Fixes the robot rotation that was failing and being out of sync when we press a unknown key.
Added borders in level 1.

We also finally decided to put a border of 22, because it becomes insane when the size is superior.

What we need now is to:
- create some basic levels
- implement a ini and yml parser for both i18n and key/level configuration.
- implement scoring (which means implement looting (see simbad's TP3 in the university intranet)
- implement collision with ennemies
- implement a system of lives
- implement a gameover
- implement level switching